% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIMSpredictor.R
\name{CV_calc_no_empirical}
\alias{CV_calc_no_empirical}
\title{Predict CV value for a given peptide sequence and Charge state pair.}
\usage{
CV_calc_no_empirical(Sequence, Charge)
}
\arguments{
\item{Sequence}{A string contains the sequence you want to predict CV value}

\item{Charge}{Charge state for the peptide sequence you are trying to predict the CV value}
}
\value{
recomended CV value without empirical information
}
\description{
Predict CV value for a given peptide sequence and Charge state pair.
}
