% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIMSpredictor.R
\name{CV_calc_with_empirical}
\alias{CV_calc_with_empirical}
\title{Predict CV value for a given peptide sequence and Charge state pair, set a empirical value to provide a more precise prediction.}
\usage{
CV_calc_with_empirical(Sequence, Charge, empirical_CV)
}
\arguments{
\item{Sequence}{A string contains the sequence you want to predict CV value}

\item{Charge}{Charge state for the peptide sequence you are trying to predict the CV value}

\item{emprirical_cv}{CV value obtained from prior experiments where the highest intensity was detected}
}
\value{
recomended CV value with empirical information
}
\description{
Predict CV value for a given peptide sequence and Charge state pair, set a empirical value to provide a more precise prediction.
}
